+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## [4.10.1](https://github.com/JuanSeBestia/react-native-wifi-reborn/compare/v4.10.0...v4.10.1) (2023-11-06)
+
+
+### Bug Fixes
+
+* **npm:** remove engines from package.json ([8bd1385](https://github.com/JuanSeBestia/react-native-wifi-reborn/commit/8bd13854ad02350c7a2477279e551a892062ab76))
+
+# [4.10.0](https://github.com/JuanSeBestia/react-native-wifi-reborn/compare/v4.9.0...v4.10.0) (2023-10-27)
+
+
+### Features
+
+* **android:** 🌟 add support for React Native 0.73 ([#312](https://github.com/JuanSeBestia/react-native-wifi-reborn/issues/312)) ([451b92c](https://github.com/JuanSeBestia/react-native-wifi-reborn/commit/451b92c53a0e173cd48f95b38f0bcaba2c5b63a8))
+
+# [4.9.0](https://github.com/JuanSeBestia/react-native-wifi-reborn/compare/v4.8.3...v4.9.0) (2023-10-27)
+
+
+### Bug Fixes
+
+* **Android:** Open the Wifi settings panel on Android 10 and above. ([#319](https://github.com/JuanSeBestia/react-native-wifi-reborn/issues/319)) ([bd89902](https://github.com/JuanSeBestia/react-native-wifi-reborn/commit/bd89902d10a22ad3fc3a5f81534d6a817fae5cc6))
+
+
+### Features
+
+* add with-expo example ([#314](https://github.com/JuanSeBestia/react-native-wifi-reborn/issues/314)) ([f572a81](https://github.com/JuanSeBestia/react-native-wifi-reborn/commit/f572a81f4a45f5efa77bfca0b7103d372fbada7c))
+
 ## [4.8.3](https://github.com/JuanSeBestia/react-native-wifi-reborn/compare/v4.8.2...v4.8.3) (2023-05-26)